requirements.txt
argh==0.26.1
backports.shutil-get-terminal-size==1.0.0
chardet==2.3.0
colorama==0.3.2
decorator==4.0.9
defusedxml==0.4.1
Django==1.9
html5lib==0.999
ipython==4.2.0
ipython-genutils==0.1.0
oauthlib==1.1.2
peewee==2.6.3
pexpect==4.1.0
pickleshare==0.7.2
ptyprocess==0.5.1
PyJWT==1.4.0
python-apt==0.9.3.12
python-social-auth==0.2.19
python3-openid==3.0.10
pytz==2016.4
requests==2.10.0
requests-oauthlib==0.6.1
simplegeneric==0.8.1
six==1.10.0
telepot==8.1
tmc==0.3.18
traitlets==4.2.1
unattended-upgrades==0.1
urllib3==1.16
virtualenv==15.0.2
